backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic - use existing installed versions
# pydantic==2.5.0
# pydantic-settings==2.1.0

sqlmodel==0.0.16

# Database drivers (use psycopg2-binary for PostgreSQL, sqlite3 built-in for testing)
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL
# asyncpg==0.29.0         # Uncomment for PostgreSQL async

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-multipart==0.0.6
alembic==1.13.1

# Phase 3: OpenAI Agents SDK for AI chatbot
# openai-agents-sdk>=0.1.0  # Optional - fallback mode works without it
openai>=1.0.0